This is a fix for the errors FCREPO-613 and FCREPO-609.
The operations
purgeObject
purgeDatastream
modifyDatastreamByValue
modifyDatastreamByReference

all takes a boolean parameter called force. If the change violates a data contract, the change should not be committed, unless force=true. As data contracts have not been implemented, setting force=true will have no effect for the two modify operations, but cause the two purge operations to throw GeneralExceptions. We have decided to get rid of force from the api, as we will not implement data contracts in the version originally envisioned.