changes to address FCREPO-1020, now updated for Fedora 3.5 based on v3.5 tag
Added a new sub-interface for RESTFilters that need to filter the response. Response data is only wrapped and
cached for such filters. Otherwise the response is immediate. This increases the speed of large downloads and
prevents java from consuming the heap.
SearchFilter, RISearchFilter and FindObjects filter/handler were the classes that got the additional
ResponseHandlingRESTFilter interface.
This patch does not invalidate existing configuration files. It uses the ObjectsFilter mapping to identify the
/objects servlet path as before. However, this filter is only a placeholder now. The function of matching request to
objects handlers is now delegated to another class.

Conflicts:

	fcrepo-security/fcrepo-security-pep/src/main/java/org/fcrepo/server/security/xacml/pep/rest/PEP.java
	fcrepo-security/fcrepo-security-pep/src/main/java/org/fcrepo/server/security/xacml/pep/rest/filters/ObjectsFilter.java