FCREPO-1102: URL Escaping PIDs Return 404s
When REST requests used a URL-encoded colon as the PID namespace
delimiter in the URI path, a Not Found (404) error was being
returned. This was fixed by modifying the @Path annotation's regular
expression to allow for ":", "%3A", and "%3a".

In previous releases of Fedora, this was not an issue because we
were using "{pid}" directly in the @Path annotcation string, but
with the move the CXF, we started using a regular expression.
Although JAX-RS implementations are required to automatically
URL-decode @PathParam values (unless @Encoded is specified), the
regular expression evaluation in @Path annotations occurs *before*
such decoding. At least that's the case for CXF.

https://jira.duraspace.org/browse/FCREPO-1102