Obtain a write lock before retrieving and deserializing an object
An operation in which the contents of an object are read and then written based on the data read (such as addRelationship) could result in lost updates if executed by two processes on the same object. Previously, an exception would be raised (see FCREPO-1024). This change addresses the issue by obtaining the write lock before the object is retrieved.